backend/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin~=6.7.0
flask~=3.1.0
python-dotenv~=1.0.0
google-cloud-firestore>=2.19.0